Refactor Check permission  (Add check on delete) and add Forbidden error message
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C41382

+++ Vanrise/Code/Vanrise.Security.Web/Controllers/PermissionController.cs
@@ -72,8 +72,10 @@ namespace Vanrise.Security.Web.Controllers
 
         [HttpGet]
         [Route("DeletePermissions")]
-        public Vanrise.Entities.UpdateOperationOutput<object> DeletePermissions(HolderType holderType, string holderId, EntityType entityType, string entityId)
+        public object DeletePermissions(HolderType holderType, string holderId, EntityType entityType, string entityId)
         {
+            if (!_manager.HasDeletePermissions(holderType, holderId, entityType, entityId))
+                return GetUnauthorizedResponse();
             return _manager.DeletePermissions(holderType, holderId, entityType, entityId);
         }
